pyproject.toml
[tool.poetry]
name = "dead-shop"
version = "0.1.0"
description = ""
authors = ["mimfort <mimfort21@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
django-debug-toolbar = "^4.4.2"
django-mptt = "^0.16.0"
pillow = "^10.3.0"
django = "^5.0.6"
django-elasticsearch-dsl = "^8.0"

python-dotenv = "^1.0.1"
psycopg = {extras = ["binary"], version = "^3.1.19"}
django-cors-headers = "^4.4.0"
requests = "^2.32.3"
unidecode = "^1.3.8"
gunicorn = "^22.0.0"
django-storages = {extras = ["s3"], version = "^1.14.4"}
ipdb = "^0.13.13"


[tool.poetry.group.dev.dependencies]
ruff = "^0.4.7"
requests = "^2.32.3"
pylint-django = "^2.5.5"
djlint = "^1.34.1"

[tool.ruff.lint]
fixable = ["ALL"]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
